# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BrandOut": {
        "properties": {
          "business_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "industry",
          "business_type"
        ],
        "title": "BrandOut",
        "type": "object"
      },
...
      "BusinessProfileBase": {
        "properties": {
          "advertising_budget_preference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertising Budget Preference"
          },
          "alternative_presence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternative Presence"
          },
          "business_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Type"
          },
          "buying_process": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buying Process"
          },
          "campaign_goals": {
            "items": {
              "type": "string"
            },
            "title": "Campaign Goals",
            "type": "array"
          },
          "common_objections": {
            "items": {
              "type": "string"
            },
            "title": "Common Objections",
            "type": "array"
          },
          "company_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Company Name",
            "type": "string"
          },
          "compliance_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Notes"
          },
          "core_problem_solved": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Problem Solved"
          },
          "customer_pain_points": {
            "items": {
              "type": "string"
            },
            "title": "Customer Pain Points",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "google_business_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Business Profile"
          },
          "ideal_customer_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal Customer Profile"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "marketplace_links": {
            "items": {
              "type": "string"
            },
            "title": "Marketplace Links",
            "type": "array"
          },
          "markets_served": {
            "items": {
              "type": "string"
            },
            "title": "Markets Served",
            "type": "array"
          },
          "onboarding_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Path"
          },
          "preferred_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Platforms",
            "type": "array"
          },
          "pricing_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Model"
          },
          "sensitive_topics": {
            "items": {
              "type": "string"
            },
            "title": "Sensitive Topics",
            "type": "array"
          },
          "social_links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Social Links",
            "type": "object"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "unique_value_proposition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique Value Proposition"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "weekly_ad_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Ad Volume"
          }
        },
        "required": [
          "company_name"
        ],
        "title": "BusinessProfileBase",
        "type": "object"
      },
      "BusinessProfileOut": {
        "properties": {
          "advertising_budget_preference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertising Budget Preference"
          },
          "alternative_presence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternative Presence"
          },
          "brand_id": {
            "title": "Brand Id",
            "type": "string"
          },
          "business_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Type"
          },
          "buying_process": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buying Process"
          },
          "campaign_goals": {
            "items": {
              "type": "string"
            },
            "title": "Campaign Goals",
            "type": "array"
          },
          "common_objections": {
            "items": {
              "type": "string"
            },
            "title": "Common Objections",
            "type": "array"
          },
          "company_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Company Name",
            "type": "string"
          },
          "compliance_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Notes"
          },
          "core_problem_solved": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Problem Solved"
          },
          "customer_pain_points": {
            "items": {
              "type": "string"
            },
            "title": "Customer Pain Points",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "google_business_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Business Profile"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ideal_customer_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal Customer Profile"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "marketplace_links": {
            "items": {
              "type": "string"
            },
            "title": "Marketplace Links",
            "type": "array"
          },
          "markets_served": {
            "items": {
              "type": "string"
            },
            "title": "Markets Served",
            "type": "array"
          },
          "onboarding_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Path"
          },
          "preferred_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Platforms",
            "type": "array"
          },
          "pricing_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Model"
          },
          "sensitive_topics": {
            "items": {
              "type": "string"
            },
            "title": "Sensitive Topics",
            "type": "array"
          },
          "social_links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Social Links",
            "type": "object"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "unique_value_proposition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique Value Proposition"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "weekly_ad_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Ad Volume"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "company_name",
          "id",
          "brand_id",
          "workspace_id"
        ],
        "title": "BusinessProfileOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnboardingRequest": {
        "description": "Full guided-onboarding payload. No presence field is required.",
        "properties": {
          "audiences": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Audiences",
            "type": "array"
          },
          "brand_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Brand Name",
            "type": "string"
          },
          "claims": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Claims",
            "type": "array"
          },
          "competitors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Competitors",
            "type": "array"
          },
          "preferred_source_types": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Source Types",
            "type": "array"
          },
          "products": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Products",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/BusinessProfileBase"
          }
        },
        "required": [
          "brand_name",
          "profile"
        ],
        "title": "OnboardingRequest",
        "type": "object"
      },
      "OnboardingResult": {
        "properties": {
          "brand": {
            "$ref": "#/components/schemas/BrandOut"
          },
          "business_profile": {
            "$ref": "#/components/schemas/BusinessProfileOut"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "brand",
          "business_profile",
          "workspace_id",
          "onboarding_completed"
        ],
        "title": "OnboardingResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SignalNest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/brands": {
      "get": {
        "operationId": "list_brands_api_v1_workspaces__workspace_id__brands_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BrandOut"
                  },
                  "title": "Response List Brands Api V1 Workspaces  Workspace Id  Brands Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Brands",
        "tags": [
          "brands"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/business-profile": {
      "get": {
        "operationId": "get_business_profile_api_v1_workspaces__workspace_id__business_profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Business Profile",
        "tags": [
          "brands"
        ]
      },
      "put": {
        "operationId": "update_business_profile_api_v1_workspaces__workspace_id__business_profile_put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProfileBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Business Profile",
        "tags": [
          "brands"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/onboarding": {
      "post": {
        "operationId": "onboard_api_v1_workspaces__workspace_id__onboarding_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Onboard",
        "tags": [
          "brands"
        ]
      }
    },
...
  }
}